Remove viewmodel from conversations page
- I suspected viewmodel wasn't actually doing anything any more - true
- added file with viewmodel class to the models section for later use
